fix intel NM bugs

+++ libfreeipmi/cmds/ipmi-oem-intel-node-manager-cmds.c
@@ -555,10 +555,10 @@ fill_cmd_oem_intel_node_manager_set_node_manager_policy (uint8_t domain_id,
                                                          uint8_t policy_configuration_action,
                                                          uint8_t policy_exception_actions_send_alert,
                                                          uint8_t policy_exception_actions_shutdown_system,
-                                                         uint8_t power_limit,
-                                                         uint8_t correction_time_limit,
-                                                         uint8_t policy_trigger_limit,
-                                                         uint8_t statistics_reporting_period,
+                                                         uint16_t power_limit,
+                                                         uint32_t correction_time_limit,
+                                                         uint16_t policy_trigger_limit,
+                                                         uint16_t statistics_reporting_period,
                                                          fiid_obj_t obj_cmd_rq)
 {
   if (!IPMI_OEM_INTEL_NODE_MANAGER_DOMAIN_ID_VALID (domain_id)
